Cleanup
Cleaned up unused references, added some additional events for future features/refactoring

+++ StateMachine/EndState.cs
@@ -18,7 +18,7 @@ namespace Tetris.States
         {
 
         }
-        public void Update(int currentFrame)
+        public void Update(long currentFrame)
         {
 
         }
